+++ R/s3mpi-package.r
@@ -1,6 +1,0 @@
-#' s3mpi
-#'
-#' @name s3mpi
-#' @docType package
-#' @import AWS.tools stringr digest cacher
-NULL